+++ apps/webui/src/shared/config/paths.ts
@@ -14,6 +14,7 @@ export const PATHS = {
   TEST_SWITCH: `${ROOT}/bin/test_switch.sh`,
   TEST_SWITCH_BG: `${ROOT}/bin/test_switch_bg.sh`,
   DATADIR: `${ROOT}/data`,
+  CONF_RELOAD_REQ: `${ROOT}/data/conf_reload_req`,
   MODULE_OFF_FLAG: `${ROOT}/data/module_off`,
   LOG_FILE: `${ROOT}/data/log.log`,
   LIST_SWITCH: `${ROOT}/data/list_switch`,